Long 으로 변경, 폼 제출시 session 사라지는 오류 해결해야 함

+++ the_joyful_delivery/src/main/java/DAO/PayDAO.java
@@ -20,7 +20,7 @@ public class PayDAO extends BaseDAOImpl<Payment> {
 	        psmt = con.prepareStatement(sql);
 
 	        psmt.setInt(1, dto.getIdx());
-	        psmt.setInt(2, dto.getDIdx());
+	        psmt.setLong(2, dto.getDIdx());
 
 	        int result = psmt.executeUpdate(); // ✅ 이게 정답
 	        return result == 1;
